Refactor module layout.
