+++ bellows/ezsp/v4/__init__.py
@@ -4,8 +4,8 @@ import binascii
 import logging
 from typing import Any, Dict, Tuple
 
-from bellows.ezsp.api import ProtocolHandler  # noqa: F401
-import bellows.types as t
+import bellows.ezsp.api
+import bellows.types
 
 from .commands import COMMANDS  # noqa: F401
 
@@ -13,33 +13,33 @@ EZSP_VERSION = 4
 LOGGER = logging.getLogger(__name__)
 
 
-class EZSPv4(ProtocolHandler):
+class EZSPv4(bellows.ezsp.api.ProtocolHandler):
     """EZSP Version 4 Protocol version handler."""
 
     COMMANDS = COMMANDS
-    types = t
+    types = bellows.types
 
     def _ezsp_frame(self, name: str, *args: Tuple[Any, ...]) -> bytes:
         """Serialize the named frame and data."""
         c = self.COMMANDS[name]
         frame = bytes([self._seq & 0xFF, 0, c[0]])  # Frame control. TODO.  # Frame ID
-        data = t.serialize(args, c[1])
+        data = bellows.types.serialize(args, c[1])
         return frame + data
 
     async def initialize(self, ezsp_config: Dict) -> None:
         """Initialize EmberZNet per passed configuration."""
 
         for config, value in ezsp_config.items():
-            if config in (t.EzspConfigId.CONFIG_PACKET_BUFFER_COUNT.name,):
+            if config in (bellows.types.EzspConfigId.CONFIG_PACKET_BUFFER_COUNT.name,):
                 # we want to set these last
                 continue
-            await self._cfg(t.EzspConfigId[config], value)
+            await self._cfg(self.types.EzspConfigId[config], value)
 
-        c = t.EzspConfigId
+        c = self.types.EzspConfigId
         await self._cfg(c.CONFIG_END_DEVICE_POLL_TIMEOUT, 60)
         await self._cfg(c.CONFIG_END_DEVICE_POLL_TIMEOUT_SHIFT, 8)
         await self._cfg(
-            t.EzspConfigId.CONFIG_PACKET_BUFFER_COUNT,
+            self.types.EzspConfigId.CONFIG_PACKET_BUFFER_COUNT,
             ezsp_config[c.CONFIG_PACKET_BUFFER_COUNT.name],
         )
 
@@ -64,7 +64,7 @@ class EZSPv4(ProtocolHandler):
         if sequence in self._awaiting:
             expected_id, schema, future = self._awaiting.pop(sequence)
             assert expected_id == frame_id
-            result, data = t.deserialize(data, schema)
+            result, data = bellows.types.deserialize(data, schema)
             try:
                 future.set_result(result)
             except asyncio.InvalidStateError:
@@ -76,5 +76,5 @@ class EZSPv4(ProtocolHandler):
         else:
             schema = self.COMMANDS_BY_ID[frame_id][2]
             frame_name = self.COMMANDS_BY_ID[frame_id][0]
-            result, data = t.deserialize(data, schema)
+            result, data = bellows.types.deserialize(data, schema)
             self._handle_callback(frame_name, result)
